Experimenting with the dataset
The dataset used for this project was huge (over 800 MB), so I created a different file to work with it to reduce it to something I could work with given the resources and time I had. This notebook shows the procedures I carried out to ultimately get the dataset I ended up saving and using for the final project.