Shortened, unprocessed dataset (zipped)
Due to the large size of the original dataset (600k rows) and lack of advanced computational resources, I reduced the size of the dataset to under 25k rows for my system to work with. This dataset contains roughly 24.7k entries that are all raw text. This dataset is also fairly large, so I'm uploading a zip file of it.